--- a/include/godot_cpp/core/class_db.hpp
+++ b/include/godot_cpp/core/class_db.hpp
@@ -47,7 +47,6 @@
 #include <list>
 #include <mutex>
 #include <set>
-#include <string>
 #include <unordered_map>
 #include <vector>